Fix issue with features files.
In OSGi improvments, the features file was split into features and
features-core. This causes issues with the karaf-maven-plugin as
it doesn't properly process <repository> elements in feature files.
Additionally, the plugin doesn't recoginze features-core in the
dependencies section of a pom as it only considers artifacts with
classifier=features. To workaround this, merge features and
features-core into a single file.